Add method for key validation
Why only key and not value and extra as well?

Value and extra, if need be, should be validated by the specific
adapters, as they know how/what they can store the data.
And those validations may vary.

Key on the other hand should be expected to be the same regardless
of the adapter used. Meaning, if I change the adapter, and my code
contains some stash-it usages with keys in it, the rules behind
validation for those keys should not change.